backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
autogen
numpy
pandas
scikit-learn
sqlalchemy
psycopg2-binary
faiss-cpu
